pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "huweibot"
version = "0.1.0"
description = "Physical GUI automation agent skeleton"
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
  "opencv-python",
  "numpy",
  "pyserial",
  "pydantic>=2,<3",
  "pillow",
  "pyyaml",
]

[project.optional-dependencies]
dev = [
  "fastapi",
  "uvicorn",
  "rich",
  "pytest",
]

[project.scripts]
huweibot = "huweibot.main:main"
xbot = "xbot.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["huweibot*", "xbot*"]
